# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/example/comicbookroute/fragment/HomeFragment.java
#	app/src/main/java/com/example/comicbookroute/util/BookRouteAdapter.java
#	app/src/main/res/layout/bookroute_row.xml

+++ app/src/main/java/com/example/comicbookroute/model/BookRouteDao.java
@@ -19,5 +19,6 @@ public interface BookRouteDao {
     List<BookRoute> selectAllBookRoutes();
 
     @Query("SELECT * FROM BookRoute WHERE id = :id")
-    BookRoute selectBookRouteByID(int id);
+    BookRoute selectBookRouteByID(long id);
+
 }
